// src/api/models/type.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Burbot Business Driver
 * OpenAPI spec version: 0.1.0
 */

export type Type = typeof Type[keyof typeof Type];


export const Type = {
  String: 'String',
  Text: 'Text',
  Integer: 'Integer',
  BigInteger: 'BigInteger',
  Float: 'Float',
  Decimal: 'Decimal',
  Boolean: 'Boolean',
  Date: 'Date',
  DateTime: 'DateTime',
  UUID: 'UUID',
  JSON: 'JSON',
  Reference: 'Reference',
} as const;